[BEAM-7087] Moving "errors" package to its own subdirectory.
