+++ sdks/go/pkg/beam/gbk.go
@@ -17,7 +17,7 @@ package beam
 
 import (
 	"github.com/apache/beam/sdks/go/pkg/beam/core/graph"
-	"github.com/apache/beam/sdks/go/pkg/beam/internal"
+	"github.com/apache/beam/sdks/go/pkg/beam/internal/errors"
 )
 
 // GroupByKey is a PTransform that takes a PCollection of type KV<A,B>,
