+++ sdks/go/pkg/beam/internal/errors/errors.go
