+++ sdks/go/pkg/beam/pcollection.go
@@ -16,10 +16,9 @@
 package beam
 
 import (
-	"github.com/apache/beam/sdks/go/pkg/beam/internal"
-
 	"github.com/apache/beam/sdks/go/pkg/beam/core/graph"
 	"github.com/apache/beam/sdks/go/pkg/beam/core/typex"
+	"github.com/apache/beam/sdks/go/pkg/beam/internal/errors"
 )
 
 // PCollection is an immutable collection of values of type 'A', which must be
