+++ sdks/go/pkg/beam/windowing.go
@@ -18,7 +18,7 @@ package beam
 import (
 	"github.com/apache/beam/sdks/go/pkg/beam/core/graph"
 	"github.com/apache/beam/sdks/go/pkg/beam/core/graph/window"
-	"github.com/apache/beam/sdks/go/pkg/beam/internal"
+	"github.com/apache/beam/sdks/go/pkg/beam/internal/errors"
 )
 
 // WindowInto applies the windowing strategy to each element.
